+++ domain/src/main/kotlin/op/gg/jth/domain/model/remote/GeneralRepo.kt
@@ -6,7 +6,7 @@ interface GeneralRepo {
     val assist: Int
     val kdaString : String
     val cs : Int
-    val csPerMin : Int
+    val csPerMin : Float
     val contributionForKillRate : String
     val goldEarned : Int
     val totalDamageDealtToChampions : Int
